% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{opportunities}
\alias{opportunities}
\title{Opportunities for New Center Pivot Installations}
\format{
an sf dataframe.
}
\source{
Pellett 2024 Appendix I
}
\usage{
opportunities
}
\description{
Opportunities for new center pivot installations were identified based on the absence of constraints such as slopes>5\%, wetlands, conservation areas, projected future impervious surfaces, and open water. Available "unconstrained" areas of sufficient size and shape were ranked according to the distance to existing irrigated areas and the area of cultivated land overlaid by each irrigation opportunity polygon.
}
\keyword{datasets}
