% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones}
\alias{zones}
\title{Irrigation Zones}
\format{
an sf dataframe with the following columns:
\describe{
\item{id_a}{A unique ID for each polygon, integer.}
\item{Type}{Center Pivot, not center pivot, and sprinklers}
\item{year2009 - year2021}{Columns for years when observations were made (2009, 2011, 2013, 2015, 2017, 2018, 2019, and 2021). 1=irrigation observed, 0=irrigation not observed.}
\item{Acres}{Polygon acreage}
\item{County}{The county of South Carolina where the polygon is located}
\item{geometry}{A list-column of geometries (polygons)}
}
}
\source{
Pellett 2020, Sekaran and Payero 2023, Pellett 2024
}
\usage{
zones
}
\description{
Irrigation zones are spatial polygons overlaying irrigated areas.
}
\keyword{datasets}
